requirements.txt
# Core numerics and data
numpy>=1.26
pandas>=2.2
scipy>=1.12
matplotlib>=3.8
pyyaml>=6.0
tqdm>=4.66

# RL stack
gymnasium>=0.29
stable-baselines3>=2.3
torch>=2.2

# Microgrid / power system simulation
pymgrid>=1.0
pandapower>=2.14

# Hardware communication
pymodbus>=3.6

